// packages/datadog-api-client-v2/models/PartialApplicationKeyAttributes.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Attributes of a partial application key.
*/
export declare class PartialApplicationKeyAttributes {
    /**
    * Creation date of the application key.
    */
    'createdAt'?: string;
    /**
    * The last four characters of the application key.
    */
    'last4'?: string;
    /**
    * Name of the application key.
    */
    'name'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): PartialApplicationKeyAttributes;
    static serialize(data: PartialApplicationKeyAttributes): {
        [key: string]: any;
    };
    constructor();
}